audio: update algorithm libraries for opening full parameters
The beamforming version log of algorithm:
BEAMFORM-RELEASE-VERSION 2.2,RELEASE DATE 2023-08-17

This patch is based on the commit of rockit:
====
[rt_media/rt_task]: audio: add support process full parameters for AI/AO VQE

it supports AI/AO VQE parse and process full algorithm parameters.

The main changes:

Add sub-modules for AI VQE:
- AEC Delay
- BF AES/ANR/HOWLING/DOA

Add AO VQE(using config_aovqe.json), not only AO AGC(config_aoagc.json),
and sub-modules for AO VQE:
- ANR/HOWLING

Our goal is that during the debugging of the product's audio
effects, only the configuration parameters need to be updated,
and the algorithm libraries don't need to be updated frequently.

Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>
Change-Id: I26feadf1580c97cd7d9bada9530a0a711e4fa486